improve: removed unnecessary cast in memcmp

+++ ft_memcmp.c
@@ -28,5 +28,5 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 		a++;
 		b++;
 	}
-	return ((int)*a - (int)*b);
+	return (*a - *b);
 }
